chore: fix test for custom objs
