+++ tests/integration/conftest.py
@@ -1,0 +1,12 @@
+import socket
+
+import pytest
+
+
+@pytest.fixture
+def unused_tcp_port():
+    """Finds and returns an unused TCP port."""
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        s.bind(("", 0))  # Let OS pick an available port
+        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+        return s.getsockname()[1]
