+++ tests/integration/test_end_to_end.py
@@ -1,5 +1,4 @@
 import os
-import socket
 from math import isclose
 
 from keras import Model
@@ -17,15 +16,6 @@ from elephas.utils.rdd_utils import to_simple_rdd
 
 import pytest
 import numpy as np
-
-
-@pytest.fixture
-def unused_tcp_port():
-    """Finds and returns an unused TCP port."""
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        s.bind(("", 0))  # Let OS pick an available port
-        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        return s.getsockname()[1]
 
 
 COMBINATIONS = [
